[김동균] 상품 정보 조회(이름, 가격, 할인율,이미지, 브랜드, 좋아요 수, 평점)

+++ src/main/java/com/avengers/musinsa/domain/brand/entity/Brand.java
@@ -4,11 +4,11 @@ import lombok.Getter;
 
 @Getter
 public class Brand {
-    private Integer brandId;
+    private Long brandId;
     private String nameKr;
     private String nameEn;
     private String brandImage;
-    private Integer brandLikes;
+    private int brandLikes;
     private String brandInfo;
     private String brandBorn;
     private String nameKrFirstInitial;
